Fixed varint64 parser's slow path
The slow path of the varint64 parser would truncate values to 32 bits
because of a missing typecast.  That's been fixed.  We now also have
more test cases (for both varint32 and varint64) that force the parser
to use the slow path — by sending in the data in two chunks.

+++ src/protobuf/varint64.c
@@ -109,7 +109,7 @@ varint64_process_bytes(push_parser_t *parser,
 
             PUSH_DEBUG_MSG("varint64:   byte = 0x%02"PRIx8"\n", *buf);
 
-            callback->value |= (*buf & 0x7F) << shift;
+            callback->value |= ((uint64_t) (*buf & 0x7F)) << shift;
             shift += 7;
 
             callback->bytes_processed++;
